Last Update
Everything checked and ready to send for evaluation

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: joaperei <joaperei@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/10 11:59:33 by joaperei          #+#    #+#             */
-/*   Updated: 2021/02/13 19:54:20 by joaperei         ###   ########.fr       */
+/*   Updated: 2021/02/14 17:31:53 by joaperei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *str1, const char *str2, int num)
+#include "libft.h"
+
+int	ft_strncmp(const char *str1, const char *str2, size_t num)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str1[i] == str2[i] && i < num)
